fix : racing manager 분리
