+++ package.json
@@ -6,7 +6,6 @@
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-import": "^2.25.3",
     "eslint-plugin-prettier": "^4.2.1",
-    "jsdoc": "^4.0.0",
     "prettier": "^2.8.0"
   }
 }
