+++ src/js/event/eventHandler.js
@@ -1,4 +1,6 @@
-import racingManager from '../model/racingManager.js';
+import carName from '../service/carName.js';
+import racingCar from '../service/racingCar.js';
+import racingResult from '../service/racingResult.js';
 import {
   showTrialForm,
   focusNameInput,
@@ -20,10 +22,10 @@ import car from '../model/Car.js';
 export const handleSubmitName = event => {
   event.preventDefault();
   const { value: nameInputValue } = event.target.elements['car-name'];
-  const trimmedValue = racingManager.trimNames(nameInputValue);
-  const names = racingManager.splitName(trimmedValue);
+  const trimmedValue = carName.trimNames(nameInputValue);
+  const names = carName.splitName(trimmedValue);
 
-  if (!racingManager.isValidNames(names)) {
+  if (!carName.isValidNames(names)) {
     focusNameInput();
     return;
   }
@@ -36,7 +38,7 @@ export const handleSubmitTrialCount = event => {
   event.preventDefault();
   const trialCount = Number(event.target.elements['trial-count'].value);
 
-  if (!racingManager.isValidTrialCount(trialCount)) {
+  if (!racingCar.isValidTrialCount(trialCount)) {
     focusTrialInput();
     return;
   }
@@ -44,10 +46,10 @@ export const handleSubmitTrialCount = event => {
   car.trialCount = trialCount;
   toggleDisabledTrial();
 
-  const gameResult = racingManager.generateGame();
+  const gameResult = racingCar.generateGame();
   car.gameResult = gameResult;
 
-  const winners = racingManager.getWinner();
+  const winners = racingResult.getWinner();
   car.winners = winners;
 
   updateResult(gameResult);
@@ -57,7 +59,7 @@ export const handleSubmitTrialCount = event => {
 };
 
 export const handleClickReset = () => {
-  racingManager.resetAll();
+  racingCar.resetAll();
   resetNameForm();
   toggleDisabledName();
